LayoutTests/imported/w3c/web-platform-tests/css/css-anchor-position/position-try-initial-transition-expected.txt


FAIL No transition for initial style with @position-try assert_equals: expected "50px" but got "10px"
